api-processing/requirements.txt
fastapi
uvicorn
python-dotenv
azure-identity
azure-storage-blob
azure-servicebus
opentelemetry-api
opentelemetry-sdk
azure-monitor-opentelemetry
azure-core-tracing-opentelemetry
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
python-multipart
markitdown
pydantic